Create mining.py
final assignment computer vision. 
need to decide on line color for starting zone. Right now it's pink which is detected instead of the mining board marker. If the starting line is a different color than pink then after the robot has the ice it can re-use the same code that got it to the mining area to get it to the starting area